refactor(frontend): use services index import
