+++ frontend/__tests__/.server/domain/services/marital-status.service.test.ts
@@ -5,7 +5,7 @@ import type { ServerConfig } from '~/.server/configs/server.config';
 import type { MaritalStatusDto } from '~/.server/domain/dtos';
 import type { MaritalStatusDtoMapper } from '~/.server/domain/mappers';
 import type { MaritalStatusRepository } from '~/.server/domain/repositories/marital-status.repository';
-import { MaritalStatusServiceImpl } from '~/.server/domain/services/marital-status.service';
+import { MaritalStatusServiceImpl } from '~/.server/domain/services';
 import type { LogFactory, Logger } from '~/.server/factories/log.factory';
 
 vi.mock('moize');
